tweaked analog input to work properly Still need to look at a clean way of truncating values so its not always different
fixed bug in toggleoutput for when an output is defined in HogKeys but not in
lua export

+++ src/HogKeys IO/Inputs/AnalogInput.cs
@@ -81,7 +81,7 @@ namespace net.willshouse.HogKeys.IO.Inputs
 
         public override string generateState(int[] pokeysValues)
         {
-            position = (pokeysValues[index] - calMin) / calRange;
+            position = (double)(pokeysValues[index] - calMin) / (double)calRange;
             return (minValue + (position * spread)).ToString();
         }
